浮動小数点型を保存→回復でもとの値を回復できない不具合対応 　　Jsonで保存桁数として適切でない値を選んでいた。 　　MinGWに最適化不具合があり、saveTestArray()とloadTestArray()の動作異常が起こる。 　　何もしないcutOptimize()関数を呼び出して対処できた。

+++ source/reference_and_test/version/ver1/test_basic_process.cpp
@@ -235,17 +235,17 @@ void saveBasicProcess(tSerializer& iSerializer)
     saveTestArray<float, 5>
     (
         iSerializer,
-        [](std::size_t i){return static_cast<float>(i)+0.1F;}
+        [](std::size_t i){return static_cast<float>(i+0.1);}
     );
     saveTestArray<double, 6>
     (
         iSerializer,
-        [](std::size_t i){return static_cast<double>(i)+0.2;}
+        [](std::size_t i){return i+0.2;}
     );
     saveTestArray<long double, 7>
     (
         iSerializer,
-        [](std::size_t i){return static_cast<long double>(i)+0.3L;}
+        [](std::size_t i){return static_cast<long double>(i+0.3);}
     );
 
     // 文字列型
@@ -546,17 +546,17 @@ void loadBasicProcess(tSerializer& iSerializer)
     loadTestArray<float, 5>
     (
         iSerializer,
-        [](std::size_t i){return static_cast<float>(i)+0.1F;}
+        [](std::size_t i){return static_cast<float>(i+0.1);}
     );
     loadTestArray<double, 6>
     (
         iSerializer,
-        [](std::size_t i){return static_cast<double>(i)+0.2;}
+        [](std::size_t i){return i+0.2;}
     );
     loadTestArray<long double, 7>
     (
         iSerializer,
-        [](std::size_t i){return static_cast<long double>(i)+0.3L;}
+        [](std::size_t i){return static_cast<long double>(i+0.3);}
     );
 
     // 文字列型
